vqa_project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.10
isort==6.0.1
joblib==1.5.1
loguru==0.7.3
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.5.0
multiprocess==0.70.16
mypy_extensions==1.1.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
pyarrow==20.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
-e git+https://github.com/DukeFens/vqa_project.git@d65bb1081858995478c9b0dd76d3c2d4e25198e0#egg=vqa_project&subdirectory=vqa_project
xxhash==3.5.0
yarl==1.20.1
